+++ extensions/prisma-docs-search/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Prisma Docs Search Changelog
 
+## [Windows Support] - {PR_MERGE_DATE}
+
+- Add Windows support
+- Update dependencies
+
 ## [Added Prisma Docs Search] - 2024-10-04
 
 - A cool new way to query Prisma docs